+++ lib/core/data/event/repository/event_repository_impl.dart
@@ -5,6 +5,7 @@ import 'package:app/core/data/event/dtos/event_dtos.dart';
 import 'package:app/core/data/event/dtos/event_join_request_dto/event_join_request_dto.dart';
 import 'package:app/core/data/event/dtos/event_rsvp_dto/event_rsvp_dto.dart';
 import 'package:app/core/data/event/dtos/event_story_dto/event_story_dto.dart';
+import 'package:app/core/data/event/dtos/generate_event_invitation_url_response_dto/generate_event_invitation_url_response_dto.dart';
 import 'package:app/core/data/event/gql/event_mutation.dart';
 import 'package:app/core/data/event/gql/event_query.dart';
 import 'package:app/core/domain/event/entities/event.dart';
@@ -15,6 +16,7 @@ import 'package:app/core/domain/event/entities/event_cohost_request.dart';
 import 'package:app/core/domain/event/entities/event_join_request.dart';
 import 'package:app/core/domain/event/entities/event_rsvp.dart';
 import 'package:app/core/domain/event/entities/event_story.dart';
+import 'package:app/core/domain/event/entities/generate_event_invitation_url_response.dart';
 import 'package:app/core/domain/event/event_repository.dart';
 import 'package:app/core/domain/event/input/accept_event_input/accept_event_input.dart';
 import 'package:app/core/domain/event/input/get_event_detail_input.dart';
@@ -33,6 +35,7 @@ import 'package:app/graphql/backend/event/mutation/cancel_event.graphql.dart';
 import 'package:app/graphql/backend/event/mutation/update_event_story_image.graphql.dart';
 import 'package:app/graphql/backend/event/mutation/create_event_join_request.graphql.dart';
 import 'package:app/graphql/backend/event/query/export_event_tickets.graphql.dart';
+import 'package:app/graphql/backend/event/query/generate_event_invitation_url.graphql.dart';
 import 'package:app/graphql/backend/event/query/get_event_application_answers.graphql.dart';
 import 'package:app/graphql/backend/event/query/get_event_cohost_requests.graphql.dart';
 import 'package:app/graphql/backend/event/query/get_event_checkins.graphql.dart';
@@ -591,6 +594,29 @@ class EventRepositoryImpl implements EventRepository {
       Event.fromDto(
         EventDto.fromJson(
           result.parsedData!.cancelEvent.toJson(),
+        ),
+      ),
+    );
+  }
+
+  @override
+  Future<Either<Failure, GenerateEventInvitationUrlResponse>>
+      generateEventInvitationUrl({
+    required String eventId,
+  }) async {
+    final result = await client.query$GenerateEventInvitationUrl(
+      Options$Query$GenerateEventInvitationUrl(
+        variables: Variables$Query$GenerateEventInvitationUrl(event: eventId),
+        fetchPolicy: FetchPolicy.networkOnly,
+      ),
+    );
+    if (result.hasException || result.parsedData == null) {
+      return Left(Failure.withGqlException(result.exception));
+    }
+    return Right(
+      GenerateEventInvitationUrlResponse.fromDto(
+        GenerateEventInvitationUrlResponseDto.fromJson(
+          result.parsedData!.generateEventInvitationUrl?.toJson() ?? {},
         ),
       ),
     );