+++ lib/core/presentation/pages/event/event_detail_page/guest_event_detail_page/widgets/guest_event_detail_dashboard.dart
@@ -1,3 +1,4 @@
+import 'package:app/core/application/event_tickets/generate_event_invitation_url_bloc/generate_event_invitation_url_bloc.dart';
 import 'package:app/core/application/event_tickets/get_my_tickets_bloc/get_my_tickets_bloc.dart';
 import 'package:app/core/domain/event/entities/event.dart';
 import 'package:app/core/presentation/widgets/event/event_dashboard_item.dart';
@@ -98,12 +99,21 @@ class GuestEventDetailDashboard extends StatelessWidget {
           },
         ),
         SizedBox(width: 10.w),
-        EventDashboardItem(
-          icon: Assets.icons.icShareGradient
-              .svg(width: Sizing.small, height: Sizing.small),
-          onTap: () {
-            Vibrate.feedback(FeedbackType.light);
-            ShareUtils.shareEvent(event);
+        BlocBuilder<GenerateEventInvitationUrlBloc,
+            GenerateEventInvitationUrlState>(
+          builder: (context, state) {
+            final invitationUrl = state.whenOrNull(
+              success: (invitationUrl) => invitationUrl,
+            );
+            return EventDashboardItem(
+              loading: invitationUrl == null,
+              icon: Assets.icons.icShareGradient
+                  .svg(width: Sizing.small, height: Sizing.small),
+              onTap: () {
+                Vibrate.feedback(FeedbackType.light);
+                ShareUtils.shareEvent(context, event);
+              },
+            );
           },
         ),
       ],